feat: Finalize 'Git: Checkout Remote Branch' command
Finished building out the git_checkout_remote_branch command to add
the ability to checkout remote branches that are not present locally.
Added this as a new command, separate from git_checkout_branch to avoid
bloating the current feature and to allow users to choose which remote
to use. This has been an open request since 2013: Issue #29

Usage:

 1. Open the Command Pallette, start typing some sequential subset
 of "Git: Checkout Remote Branch", and select the command

 2. Select a remote from the list (if any remotes have been set;
 warns and exits immediately if none can be found)

 3. Select a branch from the list of that remote's branches that do
 not yet exist locally (if any such branches exist; warns and exits
 after selecting the remote if none can be found)

This functionality would be fairly easy to move under a previously-
defined command if necessary, but I think keeping it separate makes the
experience cleaner. And the current command name matches
"git checkout branch", so it should be easy for users to find, even
without searching for it.

Overall, I tried to keep the code as centalized as possible so as
not to touch too many different files, even though some parts of it
should probably be extracted, for example, into `sgit/helpers.py`.
Here are some additional details:

 - Ensured that only remote branches that do not exist locally are
available when using this command.

 - Removed comments and excess whitespace to more closely match the
rest of the code.

 - Finished adding existence assertions to remove the possibility of
this command failing silently.

 - Refactored the double list-comparison operation between the
list of remote branches and the list of local branches into a set union
operation to ensure essentially linear run-time (though the typical
lengths involved should be too small for that to matter much anyway.)

Resolves: #29